Remove integer type lengths in schema migrations
Lengths for integer types in SQL DDL statements are unnecessary and
meaningless. What is worse is it gives the reader/maintainer a false
sense of storage length. In addition removing these mostly ignored
length suffices from integer tyes in DDL we remove one more barrier
to being DDL conformant to supporting PostgreSQL if we want to though
there are still problems with `AUTO_INCREMENT`.

To demonstrate that these lengths are useless please see the integer
types from the MySQL v5.5 documentation:
https://dev.mysql.com/doc/refman/5.5/en/integer-types.html

And you can see a sanity check I documented in this Gist:
https://gist.github.com/mbbx6spp/77fc5c2c386ee6b3ef5c